requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
bcrypt==4.0.1
python-dotenv==1.0.0
